Add test_operators
