+++ Lib/test/test_decimal/test_operators.py
@@ -1,0 +1,310 @@
+import unittest
+import operator
+from . import (load_tests_for_base_classes, setUpModule, tearDownModule)
+
+class ArithmeticOperatorsTest:
+    '''Unit tests for all arithmetic operators, binary and unary.'''
+
+    def test_addition(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('-11.1')
+        d2 = Decimal('22.2')
+
+        #two Decimals
+        self.assertEqual(d1+d2, Decimal('11.1'))
+        self.assertEqual(d2+d1, Decimal('11.1'))
+
+        #with other type, left
+        c = d1 + 5
+        self.assertEqual(c, Decimal('-6.1'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 5 + d1
+        self.assertEqual(c, Decimal('-6.1'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 += d2
+        self.assertEqual(d1, Decimal('11.1'))
+
+        #inline with other type
+        d1 += 5
+        self.assertEqual(d1, Decimal('16.1'))
+
+    def test_subtraction(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('-11.1')
+        d2 = Decimal('22.2')
+
+        #two Decimals
+        self.assertEqual(d1-d2, Decimal('-33.3'))
+        self.assertEqual(d2-d1, Decimal('33.3'))
+
+        #with other type, left
+        c = d1 - 5
+        self.assertEqual(c, Decimal('-16.1'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 5 - d1
+        self.assertEqual(c, Decimal('16.1'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 -= d2
+        self.assertEqual(d1, Decimal('-33.3'))
+
+        #inline with other type
+        d1 -= 5
+        self.assertEqual(d1, Decimal('-38.3'))
+
+    def test_multiplication(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('-5')
+        d2 = Decimal('3')
+
+        #two Decimals
+        self.assertEqual(d1*d2, Decimal('-15'))
+        self.assertEqual(d2*d1, Decimal('-15'))
+
+        #with other type, left
+        c = d1 * 5
+        self.assertEqual(c, Decimal('-25'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 5 * d1
+        self.assertEqual(c, Decimal('-25'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 *= d2
+        self.assertEqual(d1, Decimal('-15'))
+
+        #inline with other type
+        d1 *= 5
+        self.assertEqual(d1, Decimal('-75'))
+
+    def test_division(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('-5')
+        d2 = Decimal('2')
+
+        #two Decimals
+        self.assertEqual(d1/d2, Decimal('-2.5'))
+        self.assertEqual(d2/d1, Decimal('-0.4'))
+
+        #with other type, left
+        c = d1 / 4
+        self.assertEqual(c, Decimal('-1.25'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 4 / d1
+        self.assertEqual(c, Decimal('-0.8'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 /= d2
+        self.assertEqual(d1, Decimal('-2.5'))
+
+        #inline with other type
+        d1 /= 4
+        self.assertEqual(d1, Decimal('-0.625'))
+
+    def test_floor_division(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('5')
+        d2 = Decimal('2')
+
+        #two Decimals
+        self.assertEqual(d1//d2, Decimal('2'))
+        self.assertEqual(d2//d1, Decimal('0'))
+
+        #with other type, left
+        c = d1 // 4
+        self.assertEqual(c, Decimal('1'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 7 // d1
+        self.assertEqual(c, Decimal('1'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 //= d2
+        self.assertEqual(d1, Decimal('2'))
+
+        #inline with other type
+        d1 //= 2
+        self.assertEqual(d1, Decimal('1'))
+
+    def test_powering(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('5')
+        d2 = Decimal('2')
+
+        #two Decimals
+        self.assertEqual(d1**d2, Decimal('25'))
+        self.assertEqual(d2**d1, Decimal('32'))
+
+        #with other type, left
+        c = d1 ** 4
+        self.assertEqual(c, Decimal('625'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 7 ** d1
+        self.assertEqual(c, Decimal('16807'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 **= d2
+        self.assertEqual(d1, Decimal('25'))
+
+        #inline with other type
+        d1 **= 4
+        self.assertEqual(d1, Decimal('390625'))
+
+    def test_module(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('5')
+        d2 = Decimal('2')
+
+        #two Decimals
+        self.assertEqual(d1%d2, Decimal('1'))
+        self.assertEqual(d2%d1, Decimal('2'))
+
+        #with other type, left
+        c = d1 % 4
+        self.assertEqual(c, Decimal('1'))
+        self.assertEqual(type(c), type(d1))
+
+        #with other type, right
+        c = 7 % d1
+        self.assertEqual(c, Decimal('2'))
+        self.assertEqual(type(c), type(d1))
+
+        #inline with decimal
+        d1 %= d2
+        self.assertEqual(d1, Decimal('1'))
+
+        #inline with other type
+        d1 %= 4
+        self.assertEqual(d1, Decimal('1'))
+
+    def test_floor_div_module(self):
+        Decimal = self.decimal.Decimal
+
+        d1 = Decimal('5')
+        d2 = Decimal('2')
+
+        #two Decimals
+        (p, q) = divmod(d1, d2)
+        self.assertEqual(p, Decimal('2'))
+        self.assertEqual(q, Decimal('1'))
+        self.assertEqual(type(p), type(d1))
+        self.assertEqual(type(q), type(d1))
+
+        #with other type, left
+        (p, q) = divmod(d1, 4)
+        self.assertEqual(p, Decimal('1'))
+        self.assertEqual(q, Decimal('1'))
+        self.assertEqual(type(p), type(d1))
+        self.assertEqual(type(q), type(d1))
+
+        #with other type, right
+        (p, q) = divmod(7, d1)
+        self.assertEqual(p, Decimal('1'))
+        self.assertEqual(q, Decimal('2'))
+        self.assertEqual(type(p), type(d1))
+        self.assertEqual(type(q), type(d1))
+
+    def test_unary_operators(self):
+        Decimal = self.decimal.Decimal
+
+        self.assertEqual(+Decimal(45), Decimal(+45))           #  +
+        self.assertEqual(-Decimal(45), Decimal(-45))           #  -
+        self.assertEqual(abs(Decimal(45)), abs(Decimal(-45)))  # abs
+
+    def test_nan_comparisons(self):
+        # comparisons involving signaling nans signal InvalidOperation
+
+        # order comparisons (<, <=, >, >=) involving only quiet nans
+        # also signal InvalidOperation
+
+        # equality comparisons (==, !=) involving only quiet nans
+        # don't signal, but return False or True respectively.
+        Decimal = self.decimal.Decimal
+        InvalidOperation = self.decimal.InvalidOperation
+        localcontext = self.decimal.localcontext
+
+        n = Decimal('NaN')
+        s = Decimal('sNaN')
+        i = Decimal('Inf')
+        f = Decimal('2')
+
+        qnan_pairs = (n, n), (n, i), (i, n), (n, f), (f, n)
+        snan_pairs = (s, n), (n, s), (s, i), (i, s), (s, f), (f, s), (s, s)
+        order_ops = operator.lt, operator.le, operator.gt, operator.ge
+        equality_ops = operator.eq, operator.ne
+
+        # results when InvalidOperation is not trapped
+        with localcontext() as ctx:
+            ctx.traps[InvalidOperation] = 0
+
+            for x, y in qnan_pairs + snan_pairs:
+                for op in order_ops + equality_ops:
+                    got = op(x, y)
+                    expected = True if op is operator.ne else False
+                    self.assertIs(expected, got,
+                                "expected {0!r} for operator.{1}({2!r}, {3!r}); "
+                                "got {4!r}".format(
+                            expected, op.__name__, x, y, got))
+
+        # repeat the above, but this time trap the InvalidOperation
+        with localcontext() as ctx:
+            ctx.traps[InvalidOperation] = 1
+
+            for x, y in qnan_pairs:
+                for op in equality_ops:
+                    got = op(x, y)
+                    expected = True if op is operator.ne else False
+                    self.assertIs(expected, got,
+                                  "expected {0!r} for "
+                                  "operator.{1}({2!r}, {3!r}); "
+                                  "got {4!r}".format(
+                            expected, op.__name__, x, y, got))
+
+            for x, y in snan_pairs:
+                for op in equality_ops:
+                    self.assertRaises(InvalidOperation, operator.eq, x, y)
+                    self.assertRaises(InvalidOperation, operator.ne, x, y)
+
+            for x, y in qnan_pairs + snan_pairs:
+                for op in order_ops:
+                    self.assertRaises(InvalidOperation, op, x, y)
+
+    def test_copy_sign(self):
+        Decimal = self.decimal.Decimal
+
+        d = Decimal(1).copy_sign(Decimal(-2))
+        self.assertEqual(Decimal(1).copy_sign(-2), d)
+        self.assertRaises(TypeError, Decimal(1).copy_sign, '-2')
+
+
+def load_tests(loader, tests, pattern):
+    return load_tests_for_base_classes(loader, tests,
+                                       [ArithmeticOperatorsTest])
+
+
+if __name__ == '__main__':
+    unittest.main()
